+++ lib/maze/option/processor.rb
@@ -23,7 +23,8 @@ module Maze
           config.enable_bugsnag = options[Maze::Option::ENABLE_BUGSNAG]
           config.tms_uri = options[Maze::Option::TMS_URI]
           config.tms_token = options[Maze::Option::TMS_TOKEN]
-          config.bugsnag_repeater_api_key = options[Maze::Option::REPEATER_API_KEY]
+          config.aspecto_repeater_api_key = options[Maze::Option::ASPECTO_REPEATER_API_KEY]
+          config.bugsnag_repeater_api_key = options[Maze::Option::BUGSNAG_REPEATER_API_KEY]
 
           # Document server options
           config.document_server_root = options[Maze::Option::DS_ROOT]